all: creating tests for core classes
