+++ package.json
@@ -19,7 +19,7 @@
     "format:fix": "prettier --write  \"src/**\" \"__tests__/**\"",
     "format": "prettier --check \"src/**\" \"__tests__/**\"",
     "lint": "eslint src __tests__",
-    "test": "nyc --reporter=text mocha --recursive \"__tests__\""
+    "test": "nyc --reporter=lcov --reporter=text --extension=.ts mocha --require ts-node/register --recursive \"__tests__/**/*.spec.ts\""
   },
   "repository": "altangent/ccxws",
   "dependencies": {
@@ -48,6 +48,7 @@
     "nyc": "^15.1.0",
     "prettier": "^2.3.0",
     "sinon": "^10.0.0",
+    "ts-node": "^10.0.0",
     "typescript": "^4.2.4"
   }
 }
